Functions/update-stamps/requirements.txt
functions-framework==3.*
firebase-admin==5.*
requests==2.*
flask==2.*
google-cloud-logging==3.*
google-cloud-firestore==2.*
google-cloud-storage==2.*